New AutoConsul::Runner::AgentProcess class
To facilitate better control of the consul agent, monitor its state,
and give more flexibility to what we do with it concurrently.

(It would be nicer to put something like this within a god or bluepill
app, but I think the various processes are too interdependent for
either to be a good choice.  Plus I want a single entry point that
stays in the foreground, ideal for running in a container).

The AgentProcess class is instantiated with the args as one wants
passed to "consul agent", and:
* It launches the thing and waits in a separate thread; we get
  immediate notification on exit.
* It attempts to verify that the agent is up by running "consul info"
  against it, which will give a non-zero exit code if the agent isn't
  up.
* It provides for "on up" callbacks and invokes them verification
  succeeds.

Next things:
* handle stopping
* handling signals
* revise launch_and_join to use the new object.