AgentProcess#while_up helper method
To facilitate concurrent operations more easily, the
AutoConsul::Runner::AgentProcess#while_up helper lets you supply
a block that you want to run in parallel while the agent is running.

When the AgentProcess moves to an :up status, your block will start
in a new thread.

When the AgentProcess moves to either :stopping or :down states,
that thread will be killed.

If you want to poll things or write out periodic metrics or
whatever, this is a reasonable way to hook into the lifecycle of
the agent.