Entry point changes for heartbeating, bugfix
Add a "-t" or "--ticks" option that, when used with the "run"
command, causes heartbeats to issue at the specified interval in
a background thread as long as the agent is in the "up" state.

This means we get a single entry point for running the full discoverable
agent.

Additionally, the "-d" option was broken and only used the default
of "/tmp/consul/state".  Now it's properly honored, which means it'll
blow up if you don't have permissions to write to the requested path.