Agent runners use optional :advertise IP.
For https://github.com/ethanrowe/auto-consul/issues/5

The AutoConsul::Runner module ::agent_runner and ::server_runner
singleton methods both allow for an option hash at the end, within
which one may provide an :advertise IP addr to use with the
consul `-advertise` option.

This allows us to advertise a different IP than that to which
we're bound.